requirements.txt
django==3.2.14
djangorestframework==3.12.4
django-cors-headers
djongo==1.3.6
PyMuPDF # Library for PDF extraction 
spaCy==3.7.5 # Library for NLP
pymongo==3.11.4
pytz #djongo's dependency
sqlparse==0.2.4

# RUN THIS COMMAND TO INSTALL THE ENGLISH LANGUAGE MODEL OF SPACY
# python -m spacy download en_core_web_sm 

# pip install --upgrade numpy
# pip uninstall thinc
# pip install thinc

